agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Kiet Ho" }]
requires-python = ">=3.12"
dependencies = [
    "uvicorn",
    "fastapi",
    "pydantic-ai-slim[ag-ui]",
    "pydantic-ai-slim[openai]",
    "python-dotenv",
    "logfire>=4.10.0",
    "fastmcp>=0.1.0",
    "httpx",
]

[tool.poetry]
name = "agent"
version = "0.1.0"
description = "Rag agentic AI"
readme = "README.md"
authors = ["Kiet Ho <htkiet@tma.com.vn>"]
package-mode = true

[tool.poetry.dependencies]
python = ">=3.12"
uvicorn = "*"
fastapi = "*"
pydantic-ai-slim = { version = "*", extras = ["ag-ui", "openai"] }
python-dotenv = "*"
logfire = ">=4.10.0"
fastmcp = ">=0.1.0"
httpx = "*"

[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"
